+++ src/arclet/alconna/formatter.py
@@ -161,11 +161,11 @@ class TextFormatter:
             if parameter.value is AllParam:
                 return f"<...{name}>"
             if not isinstance(parameter.value, BasePattern) or parameter.value.pattern != name:
-                arg += f":{parameter.value}"
+                arg += f": {parameter.value}"
             if parameter.field.display is Empty:
                 arg += " = None"
             elif parameter.field.display is not None:
-                arg += f" = {parameter.field.display} "
+                arg += f" = {parameter.field.display}"
         return f"{arg}]" if parameter.optional else f"{arg}>"
 
     def parameters(self, args: Args) -> str: